@@ -18,30 +18,36 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.17.5 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  permission_handler: ^8.1.2
-  flutter_map: ^0.13.1
-  flutter_map_location_marker: any
-  cached_network_image: ^3.0.0
-  map_elevation: ^1.1.0
-  file_picker: ^3.0.3
-  sqflite: ^2.0.0+3
-  shared_preferences: ^2.0.6
-  path: ^1.8.0
-  path_provider: ^2.0.2
-  gpx: ^2.1.0
+  flutter_map: ^1.1.1
+  flutter_map_location_marker: ^4.1.1
+  flutter_map_tile_caching:
+    git:
+      url: https://github.com/JaffaKetchup/flutter_map_tile_caching
+      ref: v5
+  latlong2: ^0.8.1
+  dynamic_color: ^1.4.0
+  geolocator: ^9.0.0
+  map_elevation: ^2.0.0
+  file_picker: ^4.6.1
+  sqflite: ^2.0.3
+  shared_preferences: ^2.0.15
+  gpx: ^2.1.1
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  google_fonts: ^3.0.1
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+
+  flutter_lints: ^2.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
