@@ -1,38 +1,35 @@
 import 'package:flutter/material.dart';
 import 'package:map_elevation/map_elevation.dart';
 import 'package:shared_preferences/shared_preferences.dart';
-import 'package:turistautak/components/map.dart';
 import 'package:turistautak/models/route.dart';
 import 'package:latlong2/latlong.dart';
-import 'package:turistautak/utils/map_downloader.dart';
-
-import 'mapview.dart';
+import 'package:turistautak/shared/components/map.dart';
 
 class RouteDetails extends StatefulWidget {
 
-  final RouteModel route;
+  final RouteModel? route;
 
-  const RouteDetails({Key key, this.route}) : super(key: key);
+  const RouteDetails({Key? key, this.route}) : super(key: key);
 
   @override
   _RouteDetailsState createState() => _RouteDetailsState();
 }
 
 class _RouteDetailsState extends State<RouteDetails> {
-  LatLng hoverPoint;
+  late LatLng hoverPoint;
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         title: Text(
-          widget.route.name,
+          widget.route!.name,
           maxLines: 2,
         ),
         actions: [
           IconButton(
             onPressed: () {
-              MyMapDownloader.downloadRegion(widget.route.getBounds(0.1), [MyMap.markedTrailsOptions, MyMap.openTopoMapOptions], context: context);
+              // MyMapDownloader.downloadRegion(widget.route.getBounds(0.1), [MyMap.markedTrailsOptions, MyMap.openTopoMapOptions], context: context);
             },
             icon: Icon(Icons.download),
           ),
@@ -41,11 +38,11 @@ class _RouteDetailsState extends State<RouteDetails> {
       body: Column(
         children: [
           Container(
-            height: MediaQuery.of(context).size.height - 290,
-            child: MyMap(
-              points: widget.route.getPoints(),
-              hoverPoint: hoverPoint,
-              bounds: widget.route.getBounds(0.1),
+            height: MediaQuery.of(context).size.height - 300,
+            child: MapComponent(
+              // points: widget.route.getPoints(),
+              // hoverPoint: hoverPoint,
+              // bounds: widget.route.getBounds(0.1),
             ),
           ),
           Container(
@@ -54,13 +51,13 @@ class _RouteDetailsState extends State<RouteDetails> {
               onNotification: (ElevationHoverNotification notification) {
                 if(notification.position != null) {
                   setState(() {
-                    hoverPoint = LatLng(notification.position.latitude, notification.position.longitude);
+                    hoverPoint = LatLng(notification.position!.latitude, notification.position!.longitude);
                   });
                 }
                 return true;
               },
               child: Elevation(
-                widget.route.getElevationPoints(),
+                widget.route!.getElevationPoints(),
                 color: Colors.green,
                 elevationGradientColors: ElevationGradientColors(
                   gt10: Colors.green,
@@ -71,19 +68,19 @@ class _RouteDetailsState extends State<RouteDetails> {
             ),
           ),
           Text(
-            '${round(widget.route.length / 1000, decimals: 2)} km',
+            '${round(widget.route!.length / 1000, decimals: 2)} km',
             style: TextStyle(
               fontSize: 30,
               fontWeight: FontWeight.bold
             ),
           ),
-          Text('▲ ${widget.route.elevationGain} m / ▼ ${widget.route.elevationLoss} m'),
+          Text('▲ ${widget.route!.elevationGain} m / ▼ ${widget.route!.elevationLoss} m'),
           ElevatedButton(
             child: Text('Indulás'),
             onPressed: () async {
               SharedPreferences prefs = await SharedPreferences.getInstance();
-              await prefs.setString('CurrentRoute', widget.route.name);
-              Navigator.pushAndRemoveUntil(context, MaterialPageRoute(builder: (context) => MapView(route: widget.route)), (route) => false);
+              await prefs.setString('CurrentRoute', widget.route!.name);
+              // Navigator.pushAndRemoveUntil(context, MaterialPageRoute(builder: (context) => MapView(route: widget.route)), (route) => false);
             },
           )
         ],
